renamed the user facing filter nodes
