+++ examples/visual_node_graph/src/system.rs
@@ -6,7 +6,7 @@ use firewheel::{
     nodes::{
         beep_saw_test::BeepSawTestNode,
         beep_test::BeepTestNode,
-        filter::const_channel_filter::ConstChannelFilterNode,
+        filter::filter::FilterNode,
         volume::{VolumeNode, VolumeNodeConfig},
         volume_pan::VolumePanNode,
         StereoToMonoNode,
@@ -65,9 +65,7 @@ impl AudioSystem {
                 }),
             ),
             NodeType::VolumePan => self.cx.add_node(VolumePanNode::default(), None),
-            NodeType::Filter => self
-                .cx
-                .add_node(ConstChannelFilterNode::<2, 16>::default(), None),
+            NodeType::Filter => self.cx.add_node(FilterNode::<2, 16>::default(), None),
         };
 
         match node_type {
